// src/project/project.interface.ts
/* eslint-disable prettier/prettier */
export interface Project {
    customerName: string;
    projectName: string;
    projectEndDate: string;
    projectStartDate: string;
    projectStatus: string;
    projectHealth: string;
    accountManager: string;
    resources: string;
    comments: string;
    priSecPM: string;
    confluenceLink: string;
    salesPerson:string;
    department: string;
    resourceName: string;
    utilization: string;

}